requirements.txt
fpdf==1.7.2
numpy==2.1.1
flask==3.0.2
flask-assets==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.1
